// pdlearning/src/web-apps/src/modules/learner/src/models/user-preference.model.ts
export enum UserPreferenceSettingKeys {
  HomeMyLearning = 'Home.MyLearning.Show',
  HomeOutstanding = 'Home.Outstanding.Show',
  HomeRecommendForU = 'Home.RecommendForU.Show',
  HomeRecommendForOrg = 'Home.RecommendForOrg.Show',
  HomeBookmark = 'Home.Bookmark.Show',
  HomeNewsFeed = 'Home.NewsFeed.Show',
  HomeShared = 'Home.Share.Show',
  HomeCalendar = 'Home.Calendar.Show'
}